Réduction des erreurs eslint
0 errors et 6 warnings maintenant, vs 12 erreurs et 8 warnings avant
