+++ web/src/components/StandistStaffHomePage.tsx
@@ -1,8 +1,8 @@
-import { useUser } from "@/lib/userContext.tsx";
 import { useTranslation } from "react-i18next";
 import { Link, useNavigate } from "@tanstack/react-router";
 import { Header } from "@/components/Header.tsx";
 import { FC, ReactNode } from "react";
+import { useUser } from "@/lib/hooks/useUser.ts";
 
 type StandistsHomeProps = {
   headerKey: string;
