+++ web/src/components/Table.tsx
@@ -3,7 +3,7 @@ import * as React from "react";
 const Table = React.forwardRef<
   HTMLTableElement,
   React.HTMLAttributes<HTMLTableElement>
->(({ className, ...props }, ref) => (
+>(({ ...props }, ref) => (
   <div className="relative w-full overflow-auto">
     <table ref={ref} className={"w-full caption-bottom text-sm"} {...props} />
   </div>
@@ -13,7 +13,7 @@ Table.displayName = "Table";
 const TableHeader = React.forwardRef<
   HTMLTableSectionElement,
   React.HTMLAttributes<HTMLTableSectionElement>
->(({ className, ...props }, ref) => (
+>(({ ...props }, ref) => (
   <thead ref={ref} className={"[&_tr]:border-b"} {...props} />
 ));
 TableHeader.displayName = "TableHeader";
@@ -21,7 +21,7 @@ TableHeader.displayName = "TableHeader";
 const TableBody = React.forwardRef<
   HTMLTableSectionElement,
   React.HTMLAttributes<HTMLTableSectionElement>
->(({ className, ...props }, ref) => (
+>(({ ...props }, ref) => (
   <tbody ref={ref} className={"[&_tr:last-child]:border-0"} {...props} />
 ));
 TableBody.displayName = "TableBody";
@@ -29,7 +29,7 @@ TableBody.displayName = "TableBody";
 const TableFooter = React.forwardRef<
   HTMLTableSectionElement,
   React.HTMLAttributes<HTMLTableSectionElement>
->(({ className, ...props }, ref) => (
+>(({ ...props }, ref) => (
   <tfoot
     ref={ref}
     className={"border-t bg-muted/50 font-medium [&>tr]:last:border-b-0"}
@@ -41,7 +41,7 @@ TableFooter.displayName = "TableFooter";
 const TableRow = React.forwardRef<
   HTMLTableRowElement,
   React.HTMLAttributes<HTMLTableRowElement>
->(({ className, ...props }, ref) => (
+>(({ ...props }, ref) => (
   <tr
     ref={ref}
     className={
@@ -55,7 +55,7 @@ TableRow.displayName = "TableRow";
 const TableHead = React.forwardRef<
   HTMLTableCellElement,
   React.ThHTMLAttributes<HTMLTableCellElement>
->(({ className, ...props }, ref) => (
+>(({ ...props }, ref) => (
   <th
     ref={ref}
     className={
@@ -69,7 +69,7 @@ TableHead.displayName = "TableHead";
 const TableCell = React.forwardRef<
   HTMLTableCellElement,
   React.TdHTMLAttributes<HTMLTableCellElement>
->(({ className, ...props }, ref) => (
+>(({ ...props }, ref) => (
   <td
     ref={ref}
     className={"p-4 align-middle [&:has([role=checkbox])]:pr-0"}
@@ -81,7 +81,7 @@ TableCell.displayName = "TableCell";
 const TableCaption = React.forwardRef<
   HTMLTableCaptionElement,
   React.HTMLAttributes<HTMLTableCaptionElement>
->(({ className, ...props }, ref) => (
+>(({ ...props }, ref) => (
   <caption
     ref={ref}
     className={"mt-4 text-sm text-muted-foreground"}
